Fix OBJ parsing bugs that caused many Blocks objects to render incorrectly.
* Faces in OBJ with missing normals or texcoods (e.g, "17//5")
were throwing a NumberFormatException. Fixed to allow faces with
missing components.

* We were incorrectly mutating a Vec3 for translation/scaling,
whereas the rest of the code assumed it was immutable. Depending
on the mesh layout, this caused some vertices to be translated/scaled
multiple times, leading to rendering distortions. Fixed by making
a copy of the Vec3 before mutating. This problem was originally
reported in:
https://github.com/googlevr/poly-sample-android/issues/4